requirements.txt
tensorflow>=2.15
numpy
pyyaml
scikit-learn
matplotlib
